requirements.txt
telethon==1.33.1
aiofiles
apscheduler
pornhub-api==0.2.0
aiohttp==3.9.3
asyncio==3.4.3
gTTS
lyricsgenius==2.0.2
dnspython==2.3.0
hachoir
heroku3
motor
pillow==9.5.0
psutil
filetype==1.2.0
wget
qrcode
future==0.18.3
search_engine_parser
speedtest
py-tgcalls==1.2.9
MukeshAPI
ntgcalls==1.1.2
gitpython
pykeyboard
python-dotenv
pyyaml==6.0.1
tgcrypto
telegraph==2.2.0
pyfiglet
nekos.py
yt-dlp
lexica-api==1.4.7
httpx==0.25.2
lyricsgenius
spotipy
SafoneAPI
youtube_search
youtube-search-python
googlesearch-python
unidecode
#git+https://github.com/joetats/youtube_search@master
#git+https://github.com/alexmercerind/youtube-search-python@main
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
git+https://github.com/AsmSafone/SafoneAPI
bing_image_downloader
google-generativeai
langdetect==1.0.9
gpytranslate
img2pdf
